Create Div2 Ranklist
The question is from SQL contest which was unrated (4oct 2023)
The goal of this SQL query is to retrieve a list of usernames from the "Results" table based on certain criteria:

It starts by focusing on the "Results" table using the FROM Results clause.

The WHERE clause finds the rows based on the "rating" column. Only rows with ratings between 1600 (inclusive) and 2000 (exclusive) are included in the result.
SELECT selects the "username" column, giving this column an alias "Div 2 Ranklist" for clear naming.
Finally, the ORDER BY clause orders the selected rows based on the "score" column in descending order. This means that the usernames will be listed in descending order of their scores