Create Subscription
This is a problem rating between 500 to 1000 and the solution is like The program calculates the number of subscriptions needed for the group. Since 6 people can share each subscription, N/6 subscriptions are sufficient. However, if the evenly distributed subscriptions do not cover any remaining friends (N % 6 != 0), an additional subscription is required. So, a is set to 1 if there are any remainders. The total cost is calculated by adding the evenly distributed subscriptions c and the additional subscription a, then multiplying it by the cost of one subscription X. This gives the minimum total cost for the group of friends to use Chef-TV.
The minimum total cost calculated in the previous step is printed as the output for the current test case.