Create Dice Game3
This is the problem from CodeChef contest starters 126(20-3-24) where the problem says :-  Chef is playing a game consisting of 
N rounds. In each round he rolls a regular six faced dice.
He starts with an initial score of 0. Let the number rolled by him in a particular round be X.
If this is not the first round and the previous number rolled was 
1,  chef will add 2⋅X to his score, otherwise, he will add X to his score. tets cases were 
Input                 Output
2
2                           13
4                           26
we clearly see the output is multiple of 7.5